Implementación manejo de excepciones

+++ src/main/java/com/example/dentalOffice/service/OdontologistServiceImpl.java
@@ -2,6 +2,7 @@ package com.example.dentalOffice.service;
 
 import com.example.dentalOffice.entity.Odontologist;
 import com.example.dentalOffice.entity.dto.OdontologistDto;
+import com.example.dentalOffice.exceptions.ResourceNotFoundException;
 import com.example.dentalOffice.repository.IOdontologistRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,12 +25,12 @@ public class OdontologistServiceImpl implements IOdontologistService {
     }
 
     @Override
-    public OdontologistDto getOdontologistById(Long id) throws Exception {
+    public OdontologistDto getOdontologistById(Long id) throws ResourceNotFoundException {
         Optional<Odontologist> found = odontologistRepository.findById(id);
         if(found.isPresent())
             return mapper.convertValue(found, OdontologistDto.class);
         else
-            throw new Exception("Odontologist Not Exist");
+            throw new ResourceNotFoundException("Odontologist Not Exist with id: "+id);
     }
 
     @Override
@@ -48,12 +49,20 @@ public class OdontologistServiceImpl implements IOdontologistService {
     }
 
     @Override
-    public void updateOdontologist(OdontologistDto odontologist) {
-        createOdontologist(odontologist);
+    public void updateOdontologist(OdontologistDto odontologist) throws ResourceNotFoundException {
+        Optional<Odontologist> found = odontologistRepository.findById(odontologist.getId());
+        if (found.isPresent())
+            createOdontologist(odontologist);
+        else
+            throw new ResourceNotFoundException("Odontologist Not Exist with id: "+odontologist.getId());
     }
 
     @Override
-    public void deleteOdontologist(Long id) {
-        odontologistRepository.deleteById(id);
+    public void deleteOdontologist(Long id) throws ResourceNotFoundException{
+        Optional<Odontologist> found = odontologistRepository.findById(id);
+        if (found.isPresent())
+            odontologistRepository.deleteById(id);
+        else
+            throw new ResourceNotFoundException("Odontologist Not Exist with id: "+id);
     }
 }
